用户回放使用定时器，取消原来的多线程轮询&优化工具易用性&修复bug&优化代码，使框架更稳定

+++ src/third/md5.c
@@ -152,7 +152,7 @@ void MD5Final(unsigned char digest[16], LPMD5_Context ctx)
 	MD5Transform(ctx->stat, (uint32_t *) ctx->buf);
 	memset(ctx->buf, 0, 56);
     } else {
-	memset(p, 0, count - 8);
+	memset(p, 0, (size_t)count - 8);
     }
 
     ((uint32_t *) ctx->buf)[14] = ctx->count[0];
